Improved and organised code in bundle menu patches
