+++ stardew-access/Patches/ChatMenuPatches.cs
@@ -1,4 +1,5 @@
-﻿using StardewValley.Menus;
+﻿using StardewValley;
+using StardewValley.Menus;
 
 namespace stardew_access.Patches
 {
